+++ Coimbra.Tests/EventSystemTests.cs
@@ -32,7 +32,7 @@ namespace Coimbra.Tests
         public void AddListener_Single()
         {
             const string log = nameof(log);
-            _eventService.AddListener(delegate(object sender, ref TestEvent testEvent)
+            _eventService.AddListener(delegate(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(log);
             });
@@ -46,12 +46,12 @@ namespace Coimbra.Tests
         {
             const string logA = nameof(logA);
             const string logB = nameof(logB);
-            _eventService.AddListener(delegate(object sender, ref TestEvent testEvent)
+            _eventService.AddListener(delegate(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(logA);
             });
 
-            _eventService.AddListener(delegate(object sender, ref TestEvent testEvent)
+            _eventService.AddListener(delegate(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(logB);
             });
@@ -70,25 +70,25 @@ namespace Coimbra.Tests
             EventHandle handle3 = new EventHandle();
             EventHandle handle4 = new EventHandle();
 
-            void callback1(object sender, ref TestEvent testEvent)
+            void callback1(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(nameof(callback1));
                 _eventService.RemoveListener(handle1);
             }
 
-            void callback2(object sender, ref TestEvent testEvent)
+            void callback2(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(nameof(callback2));
                 _eventService.RemoveListener(handle2);
             }
 
-            void callback3(object sender, ref TestEvent testEvent)
+            void callback3(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(nameof(callback3));
                 _eventService.RemoveListener(handle3);
             }
 
-            void callback4(object sender, ref TestEvent testEvent)
+            void callback4(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(nameof(callback4));
                 _eventService.RemoveListener(handle4);
@@ -122,7 +122,7 @@ namespace Coimbra.Tests
         {
             const string log = nameof(log);
 
-            static void callback(object sender, ref TestEvent testEvent)
+            static void callback(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(log);
             }
@@ -139,12 +139,12 @@ namespace Coimbra.Tests
             const string logA = nameof(logA);
             const string logB = nameof(logB);
 
-            static void callbackA(object sender, ref TestEvent testEvent)
+            static void callbackA(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(logA);
             }
 
-            static void callbackB(object sender, ref TestEvent testEvent)
+            static void callbackB(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(logB);
             }
@@ -171,12 +171,12 @@ namespace Coimbra.Tests
         {
             const string logA = nameof(logA);
             const string logB = nameof(logB);
-            _eventService.AddListener(delegate(object sender, ref TestEvent testEvent)
+            _eventService.AddListener(delegate(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(logA);
             });
 
-            _eventService.AddListener(delegate(object sender, ref TestEvent testEvent)
+            _eventService.AddListener(delegate(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(logB);
             });
@@ -191,7 +191,7 @@ namespace Coimbra.Tests
         {
             const string log = nameof(log);
             _eventService.SetEventKey<TestEvent>(new EventKey());
-            _eventService.AddListener(delegate(object sender, ref TestEvent testEvent)
+            _eventService.AddListener(delegate(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(log);
             });
@@ -207,7 +207,7 @@ namespace Coimbra.Tests
             const string log = nameof(log);
             EventKey eventKey = new EventKey();
             _eventService.SetEventKey<TestEvent>(eventKey);
-            _eventService.AddListener(delegate(object sender, ref TestEvent testEvent)
+            _eventService.AddListener(delegate(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(log);
             });
@@ -223,7 +223,7 @@ namespace Coimbra.Tests
         public void Invoke_AfterSetEventKey_AndResetEventKey()
         {
             const string log = nameof(log);
-            _eventService.AddListener(delegate(object sender, ref TestEvent testEvent)
+            _eventService.AddListener(delegate(ref EventRef<TestEvent> testEvent)
             {
                 Debug.Log(log);
             });